reformat match page
its gorgeous now
