--- app/src/main/res/layout/fragment_match_scouting_start.xml
@@ -1,145 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <TextView
-        android:id="@+id/tv_preMatch"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textSize="50sp"
-        android:textAlignment="center"
-        android:text="Pre-Match" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <TextView
-            android:id="@+id/tv_scoutInitial"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textSize="25sp"
-            android:textAlignment="center"
-            android:ems="9"
-            android:text="Scouter Initials" />
-
-        <com.google.android.material.textfield.TextInputLayout
-            android:id="@+id/til_initials"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content">
-
-            <com.google.android.material.textfield.TextInputEditText
-                android:layout_width="150dp"
-                android:layout_height="wrap_content"
-                android:hint="Initials" />
-        </com.google.android.material.textfield.TextInputLayout>
-    </LinearLayout>
-
-    <View
-        android:id="@+id/space1"
-        android:layout_width="match_parent"
-        android:layout_height="10dp" />
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <TextView
-            android:id="@+id/tv_matchNum"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textSize="25sp"
-            android:textAlignment="center"
-            android:ems="9"
-            android:text="Match #" />
-
-        <com.google.android.material.textfield.TextInputLayout
-            android:id="@+id/til_mNum"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content">
-
-            <com.google.android.material.textfield.TextInputEditText
-                android:layout_width="150dp"
-                android:layout_height="wrap_content"
-                android:hint="#" />
-        </com.google.android.material.textfield.TextInputLayout>
-    </LinearLayout>
-
-    <View
-        android:id="@+id/space2"
-        android:layout_width="match_parent"
-        android:layout_height="10dp" />
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <TextView
-            android:id="@+id/tv_robot"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textSize="25sp"
-            android:textAlignment="center"
-            android:ems="9"
-            android:text="Robot"/>
-
-        <Spinner
-            android:id="@+id/robot_spinner"
-            android:layout_width="150dp"
-            android:layout_height="50dp"
-            android:contentDescription="spinner"
-            />
-    </LinearLayout>
-
-    <View
-        android:id="@+id/space3"
-        android:layout_width="match_parent"
-        android:layout_height="10dp" />
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <TextView
-            android:id="@+id/tv_teamNum"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textSize="25sp"
-            android:textAlignment="center"
-            android:ems="9"
-            android:text="Team #" />
-
-        <com.google.android.material.textfield.TextInputLayout
-            android:id="@+id/til_tNum"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content">
-
-            <com.google.android.material.textfield.TextInputEditText
-                android:layout_width="150dp"
-                android:layout_height="wrap_content"
-                android:hint="#" />
-        </com.google.android.material.textfield.TextInputLayout>
-    </LinearLayout>
-
-    <View
-        android:id="@+id/space4"
-        android:layout_width="match_parent"
-        android:layout_height="50dp" />
-
-    <Button
-        android:id="@+id/bt_startScout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Start Scout" />
-
-
-</LinearLayout>
-
