+++ app/src/main/res/menu/activity_main_drawer.xml
@@ -11,7 +11,7 @@
             android:title="home" />
 
         <item
-            android:id="@+id/nav_match_scouting_start"
+            android:id="@+id/nav_match_scouting"
             android:icon="@drawable/icon_heart"
             android:title="match scouting" />
